type onColumnResizeGrid
